Rename the export service files.
